Moved MechanicalTest to activities.py
